requirements-build.txt
# Build-time requirements only (no PyTorch/transformers)
# These are bundled in the installer (~150MB)
# PyTorch + model downloaded on first-run (~1.5GB)
# Security Audit: 2026-03-25

fastapi==0.115.12
uvicorn[standard]==0.30.6
sse-starlette>=1.6.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
asyncpg>=0.29.0
aiosqlite==0.20.0
pydantic==2.9.0
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
xlsxwriter==3.2.0
xlrd==2.0.1
loguru==0.7.2
lxml>=4.9.0
python-jose[cryptography]>=3.4.0
PyJWT==2.12.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
httpx==0.27.0
requests>=2.32.4
python-multipart>=0.0.18
python-socketio>=5.14.0
pydantic-settings==2.5.2
email-validator==2.2.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pyyaml==6.0.2
tqdm==4.66.5
Pillow>=10.0.0
psutil>=6.0.0
pyahocorasick>=2.0.0
faiss-cpu==1.8.0
